polars_plan

Module dsl

Domain specific language for the Lazy API.

Re-exports

pub use udf::UserDefinedFunction;

pub use crate::plans::lit;

pub use function_expr::*;

pub use functions::*;

Modules

binary

dt

function_expr

functions

Functions

udf

Structs

ChainedThen

Utility struct for the when-then-otherwise expression.

ChainedWhen

Utility struct for the when-then-otherwise expression.

ExprNameNameSpace

Specialized expressions for modifying the name of existing expressions.

FieldsMapper

JoinOptions

ListNameSpace

Specialized expressions for Series of DataType::List.

RollingCovOptions

SpecialEq

Wrapper type that has special equality properties depending on the inner type specialization

StrptimeOptions

Then

Utility struct for the when-then-otherwise expression.

UnpivotArgsDSL

When

Utility struct for the when-then-otherwise expression.

Enums

AggExpr

Excluded

Expr

Expressions that can be used in various contexts.

JoinTypeOptionsIR

LazySerde

NestedType

Operator

Selector

WindowMapping

WindowType

Traits

BinaryUdfOutputField

ColumnBinaryUdf

A wrapper trait for any binary closure Fn(Column, Column) -> PolarsResult<Column>

ColumnsUdf

A wrapper trait for any closure Fn(Vec<Series>) -> PolarsResult<Series>

FunctionOutputField

RenameAliasFn

Functions

apply_multiple

Apply a function/closure over the groups of multiple columns. This should only be used in a group_by aggregation.

binary_expr

Compute op(l, r) (or equivalently l op r). l and r must have types compatible with the Operator.

first

First column in a DataFrame.

last

Last column in a DataFrame.

len

Return the number of rows in the context.

map_list_multiple

Apply a function/closure over multiple columns once the logical plan get executed.

map_multiple

Apply a function/closure over multiple columns once the logical plan get executed.

nth

Nth column in a DataFrame.

ternary_expr

when

Start a when-then-otherwise expression.

Type Aliases

GetOutput

OpaqueColumnUdf