Struct polars_plan::dsl::string::StringNameSpace

pub struct StringNameSpace(_);

Specialized expressions for Series of DataType::Utf8.

Implementations

impl StringNameSpace

pub fn contains_literal(self, pat: Expr) -> Expr

Check if a string value contains a literal substring.

pub fn contains(self, pat: Expr, strict: bool) -> Expr

Check if this column of strings contains a Regex. If strict is true, then it is an error if any pat is an invalid regex, whereas if strict is false, an invalid regex will simply evaluate to false.

pub fn ends_with(self, sub: Expr) -> Expr

Check if a string value ends with the sub string.

pub fn starts_with(self, sub: Expr) -> Expr

Check if a string value starts with the sub string.

pub fn extract(self, pat: &str, group_index: usize) -> Expr

Extract a regex pattern from the a string value. If group_index is out of bounds, null is returned.

pub fn extract_all(self, pat: Expr) -> Expr

Extract each successive non-overlapping match in an individual string as an array

pub fn count_match(self, pat: &str) -> Expr

Count all successive non-overlapping regex matches.

pub fn strptime(self, dtype: DataType, options: StrptimeOptions) -> Expr

Convert a Utf8 column into a Date/Datetime/Time column.

pub fn to_date(self, options: StrptimeOptions) -> Expr

Convert a Utf8 column into a Date column.

pub fn to_datetime( self, time_unit: Option<TimeUnit>, time_zone: Option<TimeZone>, options: StrptimeOptions ) -> Expr

Convert a Utf8 column into a Datetime column.

pub fn to_time(self, options: StrptimeOptions) -> Expr

Convert a Utf8 column into a Time column.

pub fn split(self, by: &str) -> Expr

Split the string by a substring. The resulting dtype is List<Utf8>.

pub fn split_inclusive(self, by: &str) -> Expr

Split the string by a substring and keep the substring. The resulting dtype is List<Utf8>.

pub fn replace(self, pat: Expr, value: Expr, literal: bool) -> Expr

Replace values that match a regex pat with a value.

pub fn replace_n(self, pat: Expr, value: Expr, literal: bool, n: i64) -> Expr

Replace values that match a regex pat with a value.

pub fn replace_all(self, pat: Expr, value: Expr, literal: bool) -> Expr

Replace all values that match a regex pat with a value.

pub fn strip(self, matches: Option<String>) -> Expr

Remove leading and trailing characters, or whitespace if matches is None.

pub fn lstrip(self, matches: Option<String>) -> Expr

Remove leading characters, or whitespace if matches is None.

pub fn rstrip(self, matches: Option<String>) -> Expr

Remove trailing characters, or whitespace if matches is None..

pub fn to_lowercase(self) -> Expr

Convert all characters to lowercase.

pub fn to_uppercase(self) -> Expr

Convert all characters to uppercase.

pub fn str_slice(self, start: i64, length: Option<u64>) -> Expr

Slice the string values.

pub fn explode(self) -> Expr