Enum polars_plan::logical_plan::FunctionNode

pub enum FunctionNode {
    Opaque {
        function: Arc<dyn DataFrameUdf>,
        schema: Option<Arc<dyn UdfSchema>>,
        predicate_pd: bool,
        projection_pd: bool,
        streamable: bool,
        fmt_str: &'static str,
    },
    Pipeline {
        function: Arc<dyn DataFrameUdfMut>,
        schema: SchemaRef,
        original: Option<Arc<LogicalPlan>>,
    },
    Unnest {
        columns: Arc<[Arc<str>]>,
    },
    FastProjection {
        columns: Arc<[String]>,
        duplicate_check: bool,
    },
    DropNulls {
        subset: Arc<[Arc<str>]>,
    },
    Rechunk,
    Rename {
        existing: Arc<[String]>,
        new: Arc<[String]>,
        swapping: bool,
    },
    Explode {
        columns: Arc<[Arc<str>]>,
        schema: SchemaRef,
    },
    Melt {
        args: Arc<MeltArgs>,
        schema: SchemaRef,
    },
    RowCount {
        name: Arc<str>,
        schema: SchemaRef,
        offset: Option<IdxSize>,
    },
}

Variants

Opaque

Fields

function: Arc<dyn DataFrameUdf>

schema: Option<Arc<dyn UdfSchema>>

predicate_pd: bool

allow predicate pushdown optimizations

projection_pd: bool

allow projection pushdown optimizations

streamable: bool

fmt_str: &'static str

Pipeline

Fields

function: Arc<dyn DataFrameUdfMut>

schema: SchemaRef

original: Option<Arc<LogicalPlan>>

Unnest

Fields

columns: Arc<[Arc<str>]>

FastProjection

Fields

columns: Arc<[String]>

duplicate_check: bool

DropNulls

Fields

subset: Arc<[Arc<str>]>

Rechunk

Rename

Fields

existing: Arc<[String]>

new: Arc<[String]>

swapping: bool

Explode

Fields

columns: Arc<[Arc<str>]>

schema: SchemaRef

Melt

Fields

args: Arc<MeltArgs>

schema: SchemaRef

RowCount

Fields

name: Arc<str>

schema: SchemaRef

offset: Option<IdxSize>

Implementations

impl FunctionNode

pub fn is_streamable(&self) -> bool

Whether this function can run on batches of data at a time.

pub fn expands_rows(&self) -> bool

Whether this function will increase the number of rows

pub fn evaluate(&mut self, df: DataFrame) -> PolarsResult<DataFrame>