Enum polars_plan::logical_plan::LogicalPlan

pub enum LogicalPlan {

    AnonymousScan {
        function: Arc<dyn AnonymousScan>,
        schema: SchemaRef,
        predicate: Option<Expr>,
        options: AnonymousScanOptions,
    },
    Selection {
        input: Box<LogicalPlan>,
        predicate: Expr,
    },
    Cache {
        input: Box<LogicalPlan>,
        id: usize,
        count: usize,
    },
    CsvScan {
        path: PathBuf,
        schema: SchemaRef,
        options: CsvParserOptions,
        predicate: Option<Expr>,
    },
    ParquetScan {
        path: PathBuf,
        schema: SchemaRef,
        predicate: Option<Expr>,
        options: ParquetOptions,
    },
    DataFrameScan {
        df: Arc<DataFrame>,
        schema: SchemaRef,
        output_schema: Option<SchemaRef>,
        projection: Option<Arc<Vec<String>>>,
        selection: Option<Expr>,
    },
    LocalProjection {
        expr: Vec<Expr>,
        input: Box<LogicalPlan>,
        schema: SchemaRef,
    },
    Projection {
        expr: Vec<Expr>,
        input: Box<LogicalPlan>,
        schema: SchemaRef,
    },
    Aggregate {
        input: Box<LogicalPlan>,
        keys: Arc<Vec<Expr>>,
        aggs: Vec<Expr>,
        schema: SchemaRef,
        apply: Option<Arc<dyn DataFrameUdf>>,
        maintain_order: bool,
        options: GroupbyOptions,
    },
    Join {
        input_left: Box<LogicalPlan>,
        input_right: Box<LogicalPlan>,
        schema: SchemaRef,
        left_on: Vec<Expr>,
        right_on: Vec<Expr>,
        options: JoinOptions,
    },
    HStack {
        input: Box<LogicalPlan>,
        exprs: Vec<Expr>,
        schema: SchemaRef,
    },
    Distinct {
        input: Box<LogicalPlan>,
        options: DistinctOptions,
    },
    Sort {
        input: Box<LogicalPlan>,
        by_column: Vec<Expr>,
        args: SortArguments,
    },
    Explode {
        input: Box<LogicalPlan>,
        columns: Vec<String>,
        schema: SchemaRef,
    },
    Slice {
        input: Box<LogicalPlan>,
        offset: i64,
        len: IdxSize,
    },
    Melt {
        input: Box<LogicalPlan>,
        args: Arc<MeltArgs>,
        schema: SchemaRef,
    },
    MapFunction {
        input: Box<LogicalPlan>,
        function: FunctionNode,
    },
    Union {
        inputs: Vec<LogicalPlan>,
        options: UnionOptions,
    },
    Error {
        input: Box<LogicalPlan>,
        err: Arc<Mutex<Option<PolarsError>>>,
    },
    ExtContext {
        input: Box<LogicalPlan>,
        contexts: Vec<LogicalPlan>,
        schema: SchemaRef,
    },

}

Variants

AnonymousScan

Fields

function: Arc<dyn AnonymousScan>

schema: SchemaRef

predicate: Option<Expr>

options: AnonymousScanOptions

Selection

Fields

input: Box<LogicalPlan>

predicate: Expr

Filter on a boolean mask

Cache

Fields

input: Box<LogicalPlan>

id: usize

count: usize

Cache the input at this point in the LP

CsvScan

Fields

path: PathBuf

schema: SchemaRef

options: CsvParserOptions

predicate: Option<Expr>

Filters at the scan level

Scan a CSV file

ParquetScan

Fields

path: PathBuf

schema: SchemaRef

predicate: Option<Expr>

options: ParquetOptions

Scan a Parquet file

DataFrameScan

Fields

df: Arc<DataFrame>

schema: SchemaRef

output_schema: Option<SchemaRef>

projection: Option<Arc<Vec<String>>>

selection: Option<Expr>

In memory DataFrame

LocalProjection

Fields

expr: Vec<Expr>

input: Box<LogicalPlan>

schema: SchemaRef

Projection

Fields

expr: Vec<Expr>

input: Box<LogicalPlan>

schema: SchemaRef

Column selection

Aggregate

Fields

input: Box<LogicalPlan>

keys: Arc<Vec<Expr>>

aggs: Vec<Expr>

schema: SchemaRef

apply: Option<Arc<dyn DataFrameUdf>>

maintain_order: bool

options: GroupbyOptions

Groupby aggregation

Join

Fields

input_left: Box<LogicalPlan>

input_right: Box<LogicalPlan>

schema: SchemaRef

left_on: Vec<Expr>

right_on: Vec<Expr>

options: JoinOptions

Join operation

HStack

Fields

input: Box<LogicalPlan>

exprs: Vec<Expr>

schema: SchemaRef

Adding columns to the table without a Join

Distinct

Fields

input: Box<LogicalPlan>

options: DistinctOptions

Remove duplicates from the table

Sort

Fields

input: Box<LogicalPlan>

by_column: Vec<Expr>

args: SortArguments

Sort the table

Explode

Fields

input: Box<LogicalPlan>

columns: Vec<String>

schema: SchemaRef

An explode operation

Slice

Fields

input: Box<LogicalPlan>

offset: i64

len: IdxSize

Slice the table

Melt

Fields

input: Box<LogicalPlan>

args: Arc<MeltArgs>

schema: SchemaRef

A Melt operation

MapFunction

Fields

input: Box<LogicalPlan>

function: FunctionNode

A (User Defined) Function

Union

Fields

inputs: Vec<LogicalPlan>

options: UnionOptions

Error

Fields

input: Box<LogicalPlan>

err: Arc<Mutex<Option<PolarsError>>>

Catches errors and throws them later

ExtContext

Fields

input: Box<LogicalPlan>

contexts: Vec<LogicalPlan>

schema: SchemaRef

This allows expressions to access other tables

Implementations

impl LogicalPlan

pub fn dot(
&self,
acc_str: &mut String,
id: (usize, usize),
prev_node: DotNode<'_>,
id_map: &mut PlHashMap<String, String>
) -> Result

Arguments

id - (branch, id) Used to make sure that the dot boxes are distinct. branch is an id per join/union branch id is incremented by the depth traversal of the tree.

impl LogicalPlan

pub fn schema(&self) -> PolarsResult<Cow<'_, SchemaRef>>

pub fn describe(&self) -> String