Function polars_plan::logical_plan::to_alp

pub fn to_alp(
    lp: LogicalPlan,
    expr_arena: &mut Arena<AExpr>,
    lp_arena: &mut Arena<IR>,
) -> PolarsResult<Node>

converts LogicalPlan to IR it adds expressions & lps to the respective arenas as it traverses the plan finally it returns the top node of the logical plan