polars_plan

Module plans

Re-exports

pub use crate::frame::AllowedOptimizations;

pub use crate::frame::OptFlags;

Modules

expr_ir

hive

visitor

Defines different visitor patterns and for any tree.

Structs

AExprIter

AlpIter

AnonymousScanArgs

AnonymousScanOptions

CachedSchema

DropFunction

DslBuilder

EscapeLabel

Utility structure to write to a fmt::Formatter whilst escaping the output as a label name

ExprIRDisplay

ExprIter

ExprMapper

FileInfo

IRBuilder

IRDisplay

IRDotDisplay

IRPlan

IRPlanRef

Null

The literal Null

PathsDisplay

PredicatePushDown

ProjectionPushDown

ScanSourceIter

An iterator for ScanSources

ScanSourcesDisplay

SimplifyBooleanRule

SimplifyExprRule

StackOptimizer

Optimizer that uses a stack and memory arenas in favor of recursion

TypeCoercionRule

Enums

AExpr

IR expression node that is allocated in an Arena.

Context

DslFunction

DslPlan

FileScan

FunctionIR

IR

IR is a representation of DslPlan with Nodes which are allocated in an Arena In this IR the logical plan has access to the full dataset.

IRAggExpr

LiteralValue

NodeInputs

ScanSourceRef

A reference to a single item in ScanSources

ScanSources

Set of sources to scan from

StatsFunction

Constants

NULL

Traits

AnonymousScan

ArenaExprIter

ArenaLpIter

DataFrameUdf

DataFrameUdfMut

Literal

OptimizationRule

Optimize

TypedLiteral

UdfSchema

Functions

all_elementwise

can_pre_agg

Checks whether an expression can be pre-aggregated in a group-by. Note that this also must be implemented physically, so this isn’t a complete list.

can_pre_agg_exprs

is_elementwise

Checks if the top-level expression node is elementwise. If this is the case, then stack will be extended further with any nested expression nodes.

is_elementwise_rec

Recursive variant of is_elementwise

is_elementwise_rec_no_cat_cast

Recursive variant of is_elementwise that also forbids casting to categoricals. This function is used to determine if an expression evaluation can be vertically parallelized.

is_scalar_ae

lit

Create a Literal Expression from L. A literal expression behaves like a column that contains a single distinct value.

node_to_expr

converts a node from the AExpr arena to Expr

node_to_lp

converts a node from the IR arena to a LogicalPlan

node_to_lp_cloned

optimize

permits_filter_pushdown_rec

to_aexpr

converts expression to AExpr and adds it to the arena, which uses an arena (Vec) for allocation

to_alp

to_alp_impl

converts LogicalPlan to IR it adds expressions & lps to the respective arenas as it traverses the plan finally it returns the top node of the logical plan

to_expr_ir

to_expr_ir_ignore_alias

typed_lit