polars_plan::plans

Enum FunctionIR

pub enum FunctionIR {
    RowIndex {
        name: PlSmallStr,
        offset: Option<IdxSize>,
        schema: CachedSchema,
    },
    FastCount {
        sources: ScanSources,
        scan_type: FileScan,
        alias: Option<PlSmallStr>,
    },
    Unnest {
        columns: Arc<[PlSmallStr]>,
    },
    Rechunk,
    Rename {
        existing: Arc<[PlSmallStr]>,
        new: Arc<[PlSmallStr]>,
        swapping: bool,
        schema: CachedSchema,
    },
    Explode {
        columns: Arc<[PlSmallStr]>,
        schema: CachedSchema,
    },
    Opaque {
        function: Arc<dyn DataFrameUdf>,
        schema: Option<Arc<dyn UdfSchema>>,
        predicate_pd: bool,
        projection_pd: bool,
        streamable: bool,
        fmt_str: PlSmallStr,
    },
    Pipeline {
        function: Arc<Mutex<dyn DataFrameUdfMut>>,
        schema: SchemaRef,
        original: Option<Arc<IRPlan>>,
    },
}

Variants

RowIndex

Fields

name: PlSmallStr

offset: Option<IdxSize>

schema: CachedSchema

FastCount

Fields

sources: ScanSources

scan_type: FileScan

alias: Option<PlSmallStr>

Unnest

Fields

columns: Arc<[PlSmallStr]>

Rechunk

Rename

Fields

existing: Arc<[PlSmallStr]>

new: Arc<[PlSmallStr]>

swapping: bool

schema: CachedSchema

Explode

Fields

columns: Arc<[PlSmallStr]>

schema: CachedSchema

Opaque

Fields

function: Arc<dyn DataFrameUdf>

schema: Option<Arc<dyn UdfSchema>>

predicate_pd: bool

allow predicate pushdown optimizations

projection_pd: bool

allow projection pushdown optimizations

streamable: bool

fmt_str: PlSmallStr

Pipeline

Streaming engine pipeline

Fields

function: Arc<Mutex<dyn DataFrameUdfMut>>

schema: SchemaRef

original: Option<Arc<IRPlan>>

Implementations

impl FunctionIR

pub fn is_streamable(&self) -> bool

Whether this function can run on batches of data at a time.

pub fn expands_rows(&self) -> bool

Whether this function will increase the number of rows

pub fn evaluate(&self, df: DataFrame) -> PolarsResult<DataFrame>

pub fn to_streaming_lp(&self) -> Option<IRPlanRef<'_>>