polars_plan::plans

Function to_alp_impl

pub fn to_alp_impl(
    lp: DslPlan,
    ctxt: &mut DslConversionContext<'_>,
) -> PolarsResult<Node>

converts LogicalPlan to IR it adds expressions & lps to the respective arenas as it traverses the plan finally it returns the top node of the logical plan