polars_plan::plans

Module visitor

Defines different visitor patterns and for any tree.

Structs

AexprNode

IRNode

Enums

RewriteRecursion

Controls how the TreeWalker recursion should proceed for TreeWalker::rewrite.

VisitRecursion

Controls how the TreeWalker recursion should proceed for TreeWalker::visit.

Traits

RewritingVisitor

TreeWalker

An implementor of this trait decides how and in which order its nodes get traversed Implemented for crate::dsl::Expr and AexprNode.

Visitor

Type Aliases

IRNodeArena