Module polars_plan::dsl

Domain specific language for the Lazy API.

Re-exports

pub use udf::UserDefinedFunction;

pub use crate::logical_plan::lit;

pub use functions::*;

Modules

binary

consts

dt

functions

Functions

udf

Macros

map

map_as_slice

map_owned

wrap

Structs

ChainedThen

Utility struct for the when-then-otherwise expression.

ChainedWhen

Utility struct for the when-then-otherwise expression.

FieldsMapper

JoinOptions

ListNameSpace

Specialized expressions for Series of DataType::List.

RollingCovOptions

SpecialEq

Wrapper type that has special equality properties depending on the inner type specialization

StrptimeOptions

Then

Utility struct for the when-then-otherwise expression.

When

Utility struct for the when-then-otherwise expression.

Enums

AggExpr

BooleanFunction

Excluded

Expr

Expressions that can be used in various contexts. Queries consist of multiple expressions. When using the polars lazy API, don’t construct an Expr directly; instead, create one using the functions in the polars_lazy::dsl module. See that module’s docs for more info.

FunctionExpr

Operator

WindowMapping

WindowType

Traits

BinaryUdfOutputField

FunctionOutputField

RenameAliasFn

SeriesBinaryUdf

A wrapper trait for any binary closure Fn(Series, Series) -> PolarsResult<Series>

SeriesUdf

A wrapper trait for any closure Fn(Vec<Series>) -> PolarsResult<Series>

Functions

apply_multiple

Apply a function/closure over the groups of multiple columns. This should only be used in a group_by aggregation.

binary_expr

Compute op(l, r) (or equivalently l op r). l and r must have types compatible with the Operator.

count

Count expression.

first

First column in DataFrame.

last

Last column in DataFrame.

map_list_multiple

Apply a function/closure over multiple columns once the logical plan get executed.

map_multiple

Apply a function/closure over multiple columns once the logical plan get executed.

ternary_expr

when

Start a when-then-otherwise expression.

Type Aliases

GetOutput