Module polars_plan::dsl::functions

Functions

Functions on expressions that might be useful.

Structs

DatetimeArgs

Arguments used by datetime in order to produce an Expr of Datetime

DurationArgs

Arguments used by duration in order to produce an Expr of Duration

Traits

Range

Functions

all

Selects all columns. Shorthand for col("*").

all_exprs

Create a new column with the the bitwise-and of the elements in each row.

any_exprs

Create a new column with the the bitwise-or of the elements in each row.

apply_binary

Like map_binary, but used in a groupby-aggregation context.

avg

Find the mean of all the values in the column named name. Alias for mean.

cast

Casts the column given by Expr to a different type.

coalesce

Folds the expressions from left to right keeping the first non-null values.

col

Create a Column Expression based on a column name.

cols

Select multiple columns by name.

concat_list

Concat lists entries.

cov

Compute the covariance between two columns.

date_range

Create a date range from a start and stop expression.

datetime

Construct a column of Datetime from the provided DatetimeArgs.

dtype_col

Select multiple columns by dtype.

dtype_cols

Select multiple columns by dtype.

duration

Construct a column of Duration from the provided DurationArgs

fold_exprs

Accumulate over multiple columns horizontally / row wise.

is_not_null

A column which is false wherever expr is null, true elsewhere.

is_null

A column which is true wherever expr is null, false elsewhere.

map_binary

Apply a closure on the two columns that are evaluated from Expr a and Expr b.

max

Find the maximum of all the values in the column named name. Shorthand for col(name).max().

max_exprs

Create a new column with the the maximum value per row.

mean

Find the mean of all the values in the column named name. Shorthand for col(name).mean().

median

Find the median of all the values in the column named name. Shorthand for col(name).median().

min

Find the minimum of all the values in the column named name. Shorthand for col(name).min().

min_exprs

Create a new column with the the minimum value per row.

not

Negates a boolean column.

pearson_corr

Compute the pearson correlation between two columns.

quantile

Find a specific quantile of all the values in the column named name.

range

Create a range literal.

reduce_exprs

Analogous to Iterator::reduce.

repeat

Create a column of length n containing n copies of the literal value. Generally you won’t need this function, as lit(value) already represents a column containing only value whose length is automatically set to the correct number of rows.

sum

Sum all the values in the column named name. Shorthand for col(name).sum().

sum_exprs

Create a new column with the the sum of the values in each row.

time_range

Create a time range from a start and stop expression.