Enum polars_plan::logical_plan::LogicalPlan

pub enum LogicalPlan {

    Filter {
        input: Arc<LogicalPlan>,
        predicate: Expr,
    },
    Cache {
        input: Arc<LogicalPlan>,
        id: usize,
        cache_hits: u32,
    },
    Scan {
        paths: Arc<[PathBuf]>,
        file_info: FileInfo,
        predicate: Option<Expr>,
        file_options: FileScanOptions,
        scan_type: FileScan,
    },
    DataFrameScan {
        df: Arc<DataFrame>,
        schema: SchemaRef,
        output_schema: Option<SchemaRef>,
        projection: Option<Arc<Vec<String>>>,
        selection: Option<Expr>,
    },
    Select {
        expr: Vec<Expr>,
        input: Arc<LogicalPlan>,
        schema: SchemaRef,
        options: ProjectionOptions,
    },
    GroupBy {
        input: Arc<LogicalPlan>,
        keys: Arc<Vec<Expr>>,
        aggs: Vec<Expr>,
        schema: SchemaRef,
        apply: Option<Arc<dyn DataFrameUdf>>,
        maintain_order: bool,
        options: Arc<GroupbyOptions>,
    },
    Join {
        input_left: Arc<LogicalPlan>,
        input_right: Arc<LogicalPlan>,
        schema: SchemaRef,
        left_on: Vec<Expr>,
        right_on: Vec<Expr>,
        options: Arc<JoinOptions>,
    },
    HStack {
        input: Arc<LogicalPlan>,
        exprs: Vec<Expr>,
        schema: SchemaRef,
        options: ProjectionOptions,
    },
    Distinct {
        input: Arc<LogicalPlan>,
        options: DistinctOptions,
    },
    Sort {
        input: Arc<LogicalPlan>,
        by_column: Vec<Expr>,
        slice: Option<(i64, usize)>,
        sort_options: SortMultipleOptions,
    },
    Slice {
        input: Arc<LogicalPlan>,
        offset: i64,
        len: IdxSize,
    },
    MapFunction {
        input: Arc<LogicalPlan>,
        function: FunctionNode,
    },
    Union {
        inputs: Vec<LogicalPlan>,
        options: UnionOptions,
    },
    HConcat {
        inputs: Vec<LogicalPlan>,
        schema: SchemaRef,
        options: HConcatOptions,
    },
    Error {
        input: Arc<LogicalPlan>,
        err: ErrorState,
    },
    ExtContext {
        input: Arc<LogicalPlan>,
        contexts: Vec<LogicalPlan>,
        schema: SchemaRef,
    },
    Sink {
        input: Arc<LogicalPlan>,
        payload: SinkType,
    },

}

Variants

Filter

Filter on a boolean mask

Fields

input: Arc<LogicalPlan>

predicate: Expr

Cache

Cache the input at this point in the LP

Fields

input: Arc<LogicalPlan>

id: usize

cache_hits: u32

Scan

Fields

paths: Arc<[PathBuf]>

file_info: FileInfo

predicate: Option<Expr>

file_options: FileScanOptions

scan_type: FileScan

DataFrameScan

In memory DataFrame

Fields

df: Arc<DataFrame>

schema: SchemaRef

output_schema: Option<SchemaRef>

projection: Option<Arc<Vec<String>>>

selection: Option<Expr>

Select

Polars’ select operation, this can mean projection, but also full data access.

Fields

expr: Vec<Expr>

input: Arc<LogicalPlan>

schema: SchemaRef

options: ProjectionOptions

GroupBy

Groupby aggregation

Fields

input: Arc<LogicalPlan>

keys: Arc<Vec<Expr>>

aggs: Vec<Expr>

schema: SchemaRef

apply: Option<Arc<dyn DataFrameUdf>>

maintain_order: bool

options: Arc<GroupbyOptions>

Join

Join operation

Fields

input_left: Arc<LogicalPlan>

input_right: Arc<LogicalPlan>

schema: SchemaRef

left_on: Vec<Expr>

right_on: Vec<Expr>

options: Arc<JoinOptions>

HStack

Adding columns to the table without a Join

Fields

input: Arc<LogicalPlan>

exprs: Vec<Expr>

schema: SchemaRef

options: ProjectionOptions

Distinct

Remove duplicates from the table

Fields

input: Arc<LogicalPlan>

options: DistinctOptions

Sort

Sort the table

Fields

input: Arc<LogicalPlan>

by_column: Vec<Expr>

slice: Option<(i64, usize)>

sort_options: SortMultipleOptions

Slice

Slice the table

Fields

input: Arc<LogicalPlan>

offset: i64

len: IdxSize

MapFunction

A (User Defined) Function

Fields

input: Arc<LogicalPlan>

function: FunctionNode

Union

Fields

inputs: Vec<LogicalPlan>

options: UnionOptions

HConcat

Horizontal concatenation of multiple plans

Fields

inputs: Vec<LogicalPlan>

schema: SchemaRef

options: HConcatOptions

Error

Catches errors and throws them later

Fields

input: Arc<LogicalPlan>

err: ErrorState

ExtContext

This allows expressions to access other tables

Fields

input: Arc<LogicalPlan>

contexts: Vec<LogicalPlan>

schema: SchemaRef

Sink

Fields

input: Arc<LogicalPlan>

payload: SinkType

Implementations

impl LogicalPlan

pub fn dot( &self, acc_str: &mut String, id: (usize, usize), prev_node: DotNode<'_>, id_map: &mut PlHashMap<String, String>, ) -> Result

Arguments

id - (branch, id) Used to make sure that the dot boxes are distinct. branch is an id per join/union branch id is incremented by the depth traversal of the tree.

impl LogicalPlan

pub fn schema(&self) -> PolarsResult<Cow<'_, SchemaRef>>

impl LogicalPlan

pub fn describe(&self) -> String

pub fn describe_tree_format(&self) -> String

pub fn to_alp(self) -> PolarsResult<(Node, Arena<IR>, Arena<AExpr>)>