polars_plan

Module utils

Functions

aexpr_output_name

aexpr_to_leaf_name

aexpr_to_leaf_names

aexpr_to_leaf_names_iter

column_node_to_name

comma_delimited

Utility to write comma delimited strings

expr_output_name

output name of expr

expr_to_leaf_column_name

unpack alias(col) to name of the root column name

expr_to_leaf_column_names

This should gradually replace expr_to_root_column as this will get all names in the tree.

expr_to_leaf_column_names_iter

expressions_to_schema

Take a list of expressions and a schema and determine the output schema.

has_aexpr

has_aexpr_literal

has_aexpr_window

has_expr

Can check if an expression tree has a matching_expr. This requires a dummy expression to be created that will be used to pattern match against.

has_null

merge_schemas

Concatenate multiple schemas into one, disallowing duplicate field names