Module polars_plan::utils

Traits

PushNode

Functions

aexpr_to_leaf_names

aexpr_to_leaf_names_iter

column_delimited

Utility to write comma delimited

combine_predicates_expr

expr_to_leaf_column_name

unpack alias(col) to name of the root column name

expr_to_leaf_column_names

This should gradually replace expr_to_root_column as this will get all names in the tree.

expressions_to_schema

Take a list of expressions and a schema and determine the output schema.

has_aexpr

has_aexpr_window

has_null